requirements_041225.txt
cellpose==3.1.0
GPUtil==1.4.0
numpy==2.0.2
pyimagej==1.6.0
scyjava==1.10.1
setuptools==57.4.0
setuptools==78.1.1
scikit-image==0.24.0
tifffile==2024.12.12
ttkbootstrap==1.18.0
